Do not use both "virtual" and "override".
C++'s "override" keyword implies "virtual". When "virtual" and
"override" are both specified it means something slightly different from
just "override", which is what was intended here.

I teach the C++11 class here at Google. I asked the C++ arbitors to help
me understand this difference. They warned that it is dangerous and not
worth knowing. But that it is worth avoiding.

Test: Tested on local device -- no noticable change.
BUG=115613038

Change-Id: I372821053755fc4cbb1d001b71af7f02dffde5b2

Former-commit-id: 026d70ba45832acac08d76ec962ec00dad2b9f82