Fix GroupAlertTransferHelper logic on update.
Previously, all transferred alerts were removed when the inflation was
aborted. This could cancel transfers when the child was simply updating.
Now, this only occurs on remove. On update, the previous heads up/
ambient inflation flag is carried over and the helper determines if
the content should actually heads up after inflation.

Bug: 111809944
Test: runtest systemui, manual
Change-Id: Ic7c5414849e2bfa1978bb1c57ed314c4997b8ea2

Former-commit-id: 4b8bbda21b67498c0199e88aca33ddb3b590dd5d