Replace HashMap with ArrayMap in IMMS and InputMethodUtils
With this CL, remaining use of HashMap in InputMethodManagerService
and InputMethodUtils will be replaced with ArrayMap.

There should be no user-visible behavior change because those
instances will never be used for object enumeration hence possible
difference in enumeration order does not matter.  Note that
InputMethodInfo is always enumerated with IMMS#mMethodList.

Fix: 119839847
Test: atest CtsInputMethodTestCases CtsInputMethodServiceHostTestCases
Change-Id: Iba29e55edc22478424b7bb276b61c9ac610bbc5b

Former-commit-id: 1dd7de6afdce5a227edca5e320de275def0409b8