Rephrase progress_erasing to be generic
Re-phrase this string not to refer to the specific type of storage
the device has (e.g. "sd card" or "usb storage"). It seems the
implementation is looping through various disks anyway, so it's
unlikely to always be correct. Also, the way this was achieved
was by compiling the resources differently depending on the device.
Eliminate these strings to make the framework resources generic.

Bug: 118618261
Test: make
Change-Id: Id3ef16bb1c2bf8bac24f626d3cc0ac69b2d73eb7

Former-commit-id: eb215c0d331857578cc2dd249931d21870c44f5a