Enable airplane mode when booting in safe mode
Make it possible to boot in airplane mode when it was not turned on
before shutdown/reboot, either for situations where this is required by
regulation or to work around problems in normal boot. When safe mode is
engaged (Vol-Down), disable in two stages:
1. Set Settings.Global.AIRPLANE_MODE_ON as early as possible for
   subsystems to react to it during their initialization.
2. Trigger the public ConnectiviyService API to enable airplane mode. This
   has to be done during later stages in initialization because it requires
   sending broadcasts, so some RF leakage may still happen before airplane
   mode is enabled (time window of a few seconds).

Bluetooth is not currently disabled and needs to be fixed by another
change. WiFi and NFC are turned off alongside cellular modems.

Test: manual
Bug: 120145930
Change-Id: I1d61990a9d6f7cd046d28b3e49ecf402dd485063

Former-commit-id: 7d6c8e5ec94802359809b303765d2c0c687a4baf