Hidden API usage Landroid/graphics/Canvas;->mBitmap:Landroid/graphics/Bitmap;
Added maxTargetSdk version for Android P for access of mBitmap field
through reflection. The bitmap field is configured by application usages
only already so there is no need for apps to query the Bitmap from the
Canvas and should do so directly. Getters on Canvas is generally seen
as an anti-pattern.

Bug: 117521088
Test: none
Change-Id: Ie2318248f7682fd4e0a46121bc6e9f5c1546dcfb

Former-commit-id: 29cad18bd36e35e509ff25daf8b3ccd4eed1685a