WM: Fix Letterbox placement
The letterbox is still being placed as if there were
no surface hierarchy. It is now attached to the AppWindowToken,
and when that is not positioned at (0,0), the Letterbox is
placed at the wrong place.

Instead, we now keep track of where the Letterbox is attached
to the hierarchy and offset it accordingly.

Test: Put the phone in seascape, open a max aspect ratio activity, go home. Verify the wallpaper does not flicker though the letterbox.
Bug: 120129697
Change-Id: I16068c62c11c8a1f4814405c0340600397db3fd4

Former-commit-id: 7af9d97b1ccef21d0212f0e66957616a8ed31fb2