[Multi-user] Stop migrating legacy backup setting
The backup enabled setting BACKUP_ENABLED was deprecated since Android
N and replaced by a file containing enabled state.

Stop migrating this setting for the system user to be able to
consolidate starting the backup service for system and non-system users.

There's been several Android versions since the deprecation that the
chance we need the setting is slim (as devices shouldn't OTA from N to
current). Without the setting, the default would be backup off (absence
of file) which can then be enabled again via SUW or Settings.

Bug: 120212806
Test: 1) boot and unlock system user
2) adb shell bmgr backup [package]; adb shell bmgr restore [package]
3) atest RunFrameworksServicesRoboTests
Change-Id: I6d2b375a9400dffa9f75528d42f36b4af2e14187

Former-commit-id: 3b2e750a93816c298df8b739c56fd76487f1d18f