Give navigation audio group priority over voice audio group.
As of now the voice audio group had more priority than the navigation group. Because of which navigation audio group was not shown in the Volume Dialog. This change bumps up the nav priority so it shows in the UI.

Test: Manual
Bug: 113338258

Change-Id: I98f99560867ac7e6c3ca64a07ee0608ba6f438f6
(cherry picked from commit 3d4df8345a0e3f7cd22bd7245af5686bcf9cc48c)

Former-commit-id: 82b49040e570b080559bb247722a38a4b056e56a