Skeleton implementation of RCS APIs
This change adds the classes for RCS Storage APIs. There
is no business logic implemented yet, and the parcelable boilerplate
code doesn't really do anything.

This was needed as these APIs are meant to be interconnected, i.e. to
use an RcsPart, the app developer will need an RcsMessage, and for
that they will need the RcsThread etc.

Test: Builds fine. Tests will be added as I add business logic.

Bug: 109759350
Change-Id: I20897946dc1d9218f60274c7e3242194c7e2ac32

Former-commit-id: ec851b1f9883a3ca7663b2580c4f6065b33aedc7