Multiple resumed activities
This is the first step to connecting the visibility updates to
activity lifecycle - each time when an activity is made visible,
we check if it should be made active (resume or paused) and report
the corresponding state to client if needed.

The main logical change is in
ActivityStack#ensureActivitiesVisibleLocked. The idea is that calling
ActivityRecord#makeActiveIfNeeded() should identify what should the
state of an activity be and and perform the transition. This should
eventually reduce the number of different places where we trigger
lifecycle transitions.

Now all top activities in visible and focusable stacks should be
resumed.

Bug: 117135575
Test: atest android.server.am.lifecycle
Test: atest WmTests:ActivityRecordTests
Change-Id: I67b4768cd5c51b69bb209403e2de060fe8c9e82b

Former-commit-id: 5bca8b8c25e8a944daab643b4fbf3c51662cd385