Justify location permission usage by print spooler
Also back up historical printer usage data (used to select default
printer).

The historical printer usage contains the physical location of the
printer (determined by using the permission). The data is only used by
this app and is not sent off the device beside for backup. The backup
data is not further analyzed.

Test: Built
Change-Id: I42f562ef264d49c32ff016ca1677e76cc2a2af13

Former-commit-id: dea2735a24754ae12eba2703fe49ea4c6694eaf0