Define app-op controlling legacy storage behavior.
This app-op will be used to control if apps get a "legacy" view of
storage, which means they won't be put into an isolated sandbox.  The
current plan is to set this to "allow" for all apps installed before
the Q upgrade, to keep them working well.

Making it an app-op will give privacy-sensitive users the ability to
"revoke" this functionality.

Bug: 120287776
Test: manual
Change-Id: I21d5d53966eb26dff4893017708a9316b057aa6c

Former-commit-id: e82cbb135f3f6ec0a11afb1267d178b3a65b6002