Improve @UnsupportedAppUsage docs.
Add some words aimed at app developers, to help ensure that devs are
aware of the implications of using annotated members. Also add a link
to the public info on non-SDK interfaces, which includes a feature
request link.

Also use the term "non-SDK interface" in the docs as this is what is
used externally in the blog posts.

Test: m
Change-Id: I936e53082e308f484dfd42ba56a3575b78b6d3aa

Former-commit-id: f653d935679f433ca05d1a6ef6f2d77fb49c27f3