Restores implementation of isPresenterLocked().
ag/5710281 removed the implementation of isPresenterLocked() on
StatusBarNotificationPresenter.  This method is defined as a default
method on NotificationPresenter which by default returns false; it is
not called from anywhere in SystemUI - only in ARC - so the deleted
implementation is currently unused.  However, since the method is
present on the interface, it's better to have the correct implementation
in case it gets used in the future.  Since both known implementations of
NotificationPresenter override isPresenterLocked(), the default
implementation is removed.

Test: n/a
Change-Id: Ia5aa383fde81002f13e0d892989e7ee1bc4e5cfe

Former-commit-id: 0b8174adb83287f1a076f9c3ef23363c3a86c26f