APIs to measure and delete contributed files.
In the new isolated storage world, apps can "contribute" media that
belongs to the user, which normally means it won't be deleted when
that app is uninstalled.  However, we're anticipating that some apps
might abuse this API to preserve data the user actually wants to
delete during uninstall.

This set of changes introduces new APIs to measure and delete these
contributed media items, along with tests to verify.

Bug: 116344240
Test: atest android.provider.cts.MediaStoreTest
Change-Id: Ib740e0ea74378569572cb17640ef607aaa6baf1f

Former-commit-id: 9efa7b8020c87e8040bd5741f1a70c8c85e8a584