Handle groups and importance in cached oom_adj distribution.
The code for distributing the oom_adj level across the cached
processes now takes in to account the group and importance as
it goes through them, to keep groups together.

As part of this, increase the number of cached oom_adj slots
we have to 100, allowing more room for distributing regular
processes and space between them for applying importance.  This
gives 5 slots for importance levels; if someone uses more than
5, all of the lower ones will get stuck together at the end.

Test: atest CtsAppTestCases:ServiceTest\#testActivityServiceBindingLru
Bug: 111434506
Change-Id: Ia4ec0552b824a031c2a2d62c9e42ebf446e3e5c3

Former-commit-id: b446ce5868f356bbbcb93d93e3308d10f3f9d78a