Fix max height for settings_entity_header
If user uses <include> tag to include settings_entity_header
in layout XML, it will take up all the screen and nothing
below it is shown.

The reason is layout_height of two button views
in settings_entity_header is match_parent.
So we simply changed it to wrap_content.

Test: visual test, robotest
Change-Id: Ib73382ab26b485bf6b993fb0d308e4b943f5a0a7
Fixes: 120884132

Former-commit-id: b506b23e37060aae333b6e6f26dd45180ea09ff5