[view-compiler] Add layout validation
Layout validation walks over a layout resource and verifies that the layout only
uses features the layout compiler supports.

Currently this means we reject layouts that use any of the following tags:
  * include
  * fragment
  * merge
  * view

In the future, we will extend the layout compiler to handle these, but for now
we want to be liberal in what we reject.

Bug: 111895153
Test: atest view-compiler-tests
Change-Id: I8dca30c035a83b6763497a36fc60c68438fa1b0c

Former-commit-id: 3cbf176e7c6a96c535f56cb7d10e2de79a54cbd0