About phone page getFactoryMacAddress
Currently, the about phone page is displaying whatever the current MAC
address is, and with MAC randomization enabled by default, this MAC could
change arbitrarily and depending on the currently connect network's
settings.
Instead of showing the current MAC, we should be displaying the factory
MAC address, which is more relevant in the about phone page.

Bug: 111634904
Test: atest WifiMacAddressPreferenceControllerTest
Test: enable mac randomizations and verify that in the about phone page
the factory MAC is displayed instead of the randomized MAC.
Change-Id: Ia658ae86ee7807e226aa6f560b935af27774fcf1

Former-commit-id: 6b93d8339a26b6e5610bd9792460ba8117c88d91