Fix navigation bar flaky tests
The flakiness is because the timing issue on hasNavigationBar.
When a new display is added, onDisplayAdded is called and we may try to
add nav bar before DisplayPolicy#configure is called, which sets
mHasNavigationBar value. Therefore, we may get false from
hasNavigationBar even if a display does support system decorations.

Fixes: 120748674
Test: atest ActivityManagerMultiDisplayTests
Test: atest
ActivityManagerMultiDisplayTests#testNavBarShowingOnDisplayWithDecor *
100

Change-Id: Idb3c3d0874c15a466d179cc471ade5cea890e20e

Former-commit-id: 5bdd3e27fb04a4cbba97cccb9562f520c4f2949d