Revert "[framework] Compile secondary dex files in isolation"
This reverts commit cbb8b58e603a7ba5bc0999f12715ce5c03bf42ea [formerly 8a5a708df18c2377c7ec3cb03d406bdc9cb99bd6].
ART does not compile secondary dex files in the user process
any more. The original issue of a large overhead of providing
a class loader context is therefore not relevant any more and
we can start passing the context to installd.

Note that this also partially reverts commit
476bad6d8c56798081ad98c975f04449a4437ca5 [formerly 6dba50d63375b6738c6e1ad8a8758f181d295d8b] which restricted
visibility of methods in PackageDexUsage. The commit also
removed data collection for unsupported class loaders. The
logic in PackageDexOptimizer was adjusted accordingly.

Bug: 64530081
Bug: 111174995
Test: manual
Change-Id: Id78b5a6d8841b199c12a63a8d45d12efbcc32275

Former-commit-id: 22217f6a514079c504f9d77e23645ed04700fc09