Extracts row binding logic from NotificationEntryManager.
This change introduces the NotificationRowBinder component, which
contains the logic around inflating and binding notification row views
which previously lived in NotificationEntryManager.  Separating this out
to its own component allows us to make the dependencies from other logic
in NEM explicit, and eventually to decouple them once we're not
immediately inflating views for every notification.

Test: atest SystemUITests, manually
Change-Id: Ic74108e92abb402c7aa9524b7a7c5f504879546c

Former-commit-id: 8ba88a8eb1431fc81dd4bc70b2e64183c2eca402