Changed ContentCaptureService to receive one event at time.
The initial implementation was using a batch the events to optimize the
performance, but that can be optimize behind the scenes (i.e., we can still
send the batches in the binder, but deliver them one by one).

This change not only makes it easier for the service to use the API, but it
paves the way to implement multiple sessions (so we can buffer children events
while the parent session is not completely started yet).

Bug: 121033016
Bug: 117944706
Bug: 121051220

Test: atest CtsContentCaptureServiceTestCases

Change-Id: I713ceb998bd81733255fd3ef8d0b8d7a3fcac20c

Former-commit-id: fc24beab3e91d01388ce4b7f5b62b6d5b7c6a2c0