Add Emergency Call Routing
Some countries or carriers require some emergency numbers that must
be handled with normal call routing or emergency routing.

Do not send service types from the database
as the emergency service categories

When searching for the corresponding service categories,
try to find the one that comes from the radio indication.

Test: Treehugger
Bug: 112657134
Change-Id: I95dadd07e9155c6e402a72339f307b4d7ef32fe4

Former-commit-id: a4bd738b8b4a4402a26747d6be5d9a2228a45329