Combines NotificationEntryManager listener interfaces (part 1).
This change combines the NotificationEntryManager.Callback interface
with the NotificationEntryListener interface.

Future changes will reduce duplication between the methods in these
interfaces, move the previous Callback instance into the list of
listeners instead of having its own special field, and introduce more
listeners to replace logic that's proactively invoked in
NotificationEntryManager, but this first step is just a pure combination
of the two interfaces.

Test: atest SystemUITests, manual
Change-Id: Ifafcee697e7dd35e0aaae32f677ef1582f8d5920

Former-commit-id: 26bc59b506a19e4f4c4b43f5d06b8e0716968ac4