Avoid killing instrumentation process while start process
The case is a test using the same package name as default home.
Launching the instrumentation of the test will cause default home to
be killed. But AM has a mechanism to restart the default home and that
will cause the instrumentation process to be killed.

This patch add a condition check to avoid killing instrumentation
process while starting process.

Test: atest CtsAppTestCases:ServiceTest
Bug: 120107272
Change-Id: I6da0c4ca355aa324ed76571c3bc47ac2d962d192

Former-commit-id: 0bd1bf628d857ecbe79778ce4265dc374e92e294