Use vector instead of VLA to avoid uninitialized values
I502b34f23d61a7346d79ff0dc378add8461d2d27 added a continue before
skip[i] was set, which left it uninitialized and caused
non-deterministic output of incident-section-gen incidentd and a
non-deterministic incidentd binary.  Use a vector instead of a
variable length array for skip so that it is always initialized.

Test: valgrind incident-section-gen incidentd
Change-Id: Iac9778dc8bbf4ec5540e5e2ffdaa8e2dd852d6cc

Former-commit-id: b9e195546ee80e46fca5e075c9b1ea069821f18a