Don't stop adding smart suggestions just because one doesn't fit.
Up until now we've added smart suggestions to a notification from a
certain list (in order). When we've reached a suggestion that doesn't
fit in the notification we've simply skipped the rest of the
suggestions. With this CL we only skip the suggestion that's too long,
and instead try adding the less-prioritized suggestions.

Example:
Assume the notification can fit suggestions 1, 2, and 4 (suggestion 3 is
too long to fit):
"Action1", "Action2", "LongestActionEvererererrer", "Action4"

Behaviour before this CL: only the first two suggestions are shown:
"Action1", "Action2"
Behaviour after this CL: actions 1, 2, and 4 are shown:
"Action1", "Action2", "Action4"

Bug: 121250868
Test: atest SystemUITests
Change-Id: Icd90eebe1514e88a9dd0d48db050325764830155

Former-commit-id: 35156758031dcf753f7d9fbe246c8097d11d6444