Moves minimization change handling into NotificationEntryListeners.
This change moves the logic in NotificationEntryManager which was
explicitly handling changes in whether a row was minimized or not after
it was reinflated into the target components through the normal listener
mechanism.  This involves VisualStabilityManager registering a
NotificationEntryListener and allows us to remove the dependency of
NotificationEntryManager on VisualStabilityManager.

Test: atest SystemUITests, manual
Change-Id: I4bfdbda49c1cdff4a78fabeb2343843edc509fb0

Former-commit-id: 5b9098dc3a4af11f2ddc1eae359e673d21349e00