Refine UI janky in split-screen when launch & close IME opened app
In split-screen mode, when launching the IME focused app on bottom stack,
there are 2 animations will be triggered to resize the bottom stack:

1) dock_bottom_enter transition animation start will trigger DividerView in
SysUi to resize stack & translate divider which animates from up to bottom.
2) stack resize animation for IME adjustment done by DockedStackDividerController
the direction is from bottom to up.

So it will possible cause screen flicking due to task moving in different way.

Set no animate to skip frames for 1) & make sure the enter transition animation
not running when starting IME adjust animation to fix this issue.

Bug: 118364779
Test: manual, in split-screen mode:
     1) Launch app (i.e. chrome app) on bottom stack.
     2) Tap EditText to show keyboard.
     3) Press home key & launch the same app again.
     4) Verify if animation goes well.

Change-Id: I60875554e4b7df3cbdd0241d29d5685a97ecc27c

Former-commit-id: bcc316ed3e29923c59da608587f9d4d03b3f302c