Extract IME list creation logic from IMMS#buildInputMethodListLocked()
This is a preparation to introcude following two methods into
InputMethodManagerService (IMMS).

  * getInputMethodListForUser(int userId)
  * getEnabledInputMethodListForUser(int UserId).

In order to implement those methods (with minimum changes), the logic
to create List<InputMethodInfo> for the given user needs to be
extracted out from IMMS#buildInputMethodListLocked() to avoid code
duplicate.

This CL itself is a mechanical refactoring hence there should be no
behavior change.

Bug: 120709962
Test: atest CtsInputMethodTestCases CtsInputMethodServiceHostTestCases
Change-Id: I7f7e3637ea44dddc1210bc2889b519f17bbcd545

Former-commit-id: 05139322632820201fd7d9a0521d1c789345ffc3