Fix drag and drop (2/3)
Drag and drop requires the touch focus to be transferred from one window to
another. Since InputWindows are now sent through SurfaceFlinger, there's a
possibility that the request to transfer touch focus will happen before the
input window is set in InputFlinger. This will cause the transfer focus window
to fail since the window to transfer focus to doesn't exist yet in
InputFlinger.

This change passes the window that the touch focus should be transferred from
through SurfaceFlinger so when the touch is transferred, InputFlinger is
guaranteed to know about the new window.

Change-Id: I64ac64c060f19ebbb5c4f0efa28bce3164899349
Fixes: 120463595
Fixes: 120493124
Test: Drag and drop view in app
Test: Drag and drop free form
Test: CrossAppDragAndDropTests

Former-commit-id: b5e316c3e43aaeca1a83be46eb4fde0be2e7d7d6