Clarify shell/root users also have DUMP permission
This is a follow up CL to my previous CL [1], which aimed to
streamline and centralize when and how caller verifications are
performed inside InputMethodManagerService#onShellCommand().

The goal of this change is just to make it clear that shell/root users
also have DUMP permission.  There should be no behavior change as long
as those users have DUMP permission because IMMS#onShellCommand() is
already locked down to accept IPCs only from these two UIDs [2].

The point is that ShellCommand#handleDefaultCommands() by default
dispatches 'dump' command to Binder#doDump() method.  Thus it would
make sense to clarify that DUMP permission is also granted to
shell/root users for completeness.

 [1]: I6fd47b5cc1e7da7222774df20247a2c69a70f45b
      bb68f23e7a07c0694ca6ab6234f338f270668e98 [formerly db25df71be4c1bcc654f69ce9a8bff7e3ef46360]
 [2]: If87189563ccaacd4f9c666bab4f9ad08a9343084
      e9e28d7210c2ff458f62c4dd40ce638b312b2d02 [formerly b8d240fa3f96b7b4ea35dd271beda789044d63ab]

Bug: 34886274
Test: 'adb shell cmd input_method dump' still work
Change-Id: If91137e0c144d90ac3046f30607e36bb957a7e0e

Former-commit-id: 8a2b96bb0a1d901beb9cc9635c1fe9a265799b36