TYPE_DATE and TYPE_DATE_TIME should be considered as the same type
To avoid confusion, we already suppress actions of the same entity type.
But from the user perspective, some entity types are really
the same thing, like TYPE_DATE and TYPE_DATE_TIME.

This is just a workaround before moving to suggestConversationActions
for smart actions. Once we have moved to it, both TYPE_DATE
and TYPE_DATE_TIME entities should be mapped to the same action type
--- OPEN_CALENDAR.

FIXES: 121200744
Test: atest EntityTypeCounterTest
Test: Try out a notification with both TYPE_DATE and TYPE_DATE_TIME
      entities, observes that both actions are not shown.

Change-Id: Id1c91bff5749c62be29a2da0cc9ed7fd8960d166

Former-commit-id: 1d07be390b106243b82954cd187c55063946cd41