Signed config: some renames for clarity.
"android.signedconfig" is not a good name, as many types of config could be
signed. Rename instead to "android.settings.global", as this reflects the
fact that they are really global settings values.

This also open the door to cleanly adding "android.settings.secure" or
other types of config in future, should that ever be necessary.

Also rename some classes to reflect this change, and the debug logging
accordingly.

Bug: 110509075
Test: atest CtsSignedConfigHostTestCases
Test: atest SignedConfigTest
Change-Id: I42e71fd7d7fc0305b7982fbcab49130dcf4dad4e

Former-commit-id: 44d7ef981aa5b64d65e9f77051c2a7f677b3c396