Fix WCDMA Level Calculation for Default CC
The default carrier config was setting the level
calculation method to an empty string, which caused
the level to be set forcibly to NONE/UNKNOWN due to
the logic failing to account for empty strings. This
patch sets the default to something sane "rssi" that
matches legacy behavior and also adds guards so that
an empty string will be considered default, and a non
empty but otherwise invalid string will also try to
use RSSI instead of simply failing to display signal
bars.

-Add protection against invalid array sizes
-Fix a bug where RSCP could use RSSI thresholds
-Add fallback for RSCP thresholds being null when the
 measurement type is set to RSCP.
-Define RSCP thresholds as constants.

Bug: 118800169
Test: atest CellSignalStrengthWcdmaTest
Change-Id: I8d6261bc5adab6bf9b0a88f7d59b259d338a49e9

Former-commit-id: 549c3ba3b53b58cbfd44940522aa9e779acce33c