Restart the application only for the bad process
The case is a test using the same package name as default home.
Starting the instrumentation of the test will kill the home process.
When AM try to restart the default home, there is a wrong logic that
cause the instrumentation process to be killled.

The problem of original logic is always set knownToBeDead as true even
the existing process is ok.

Test: atest CtsAppTestCases#ServiceTest
Bug: 120107272
Change-Id: Ic2c3f6236c3631b0e386d24750bf5952d8227ea9

Former-commit-id: 917173f5709b1042d74e21280635e3807c559af1