Fixes the problem of uninstalled app does not disappear from the list of Settings App info.
If there is also a work profile, the original way of checking whether some apps are removed
, mAppEntries < mApplications (compare the number of App entries with apps got from PMS)
becomes incorrect because mAppEntries filters out apps that are not installed for the owner
user.
In this case, the size of mAppEntries is not always equal to mApplications, make mAppEntries
not get rebuilt correctly.

Changes to compare the packages of latest mApplications with previous mApplications (before
Resume) to clear and rebuilt mAppEntries correctly.

Change-Id: Idf33de27f86628e52a5374a7c0303966db0b61ad
Fixes: 120685868
Test: Robo tests

Former-commit-id: 8776b557fa2154ea3edc081f043b9c6f1cbd6347