Window Manager Flag Migration (3/n)
Adds fake control which is an InsetsSourceControl without a leash in it.

Fake control will be sent to the client when bars are shown transiently.
So the client thinks it still have the control over the bars, but moving
the bars won't take effect.

InsetsController.show still works and dispatches the modified state to
the server. The server can then listen to that state change, and abort
transient showing mode, start a new animation, and send over the new
control with an actual leash to the client.

Bug: 118118435
Test: atest InsetsSourceProviderTest
            InsetsStateControllerTest
            WindowStateTests
Change-Id: I8db5942e9f951e2652d4a3a34c9fde2efe5b9858

Former-commit-id: a12ea56ba413a9bde19255be62cf6d12f228444a