Exclude IWLAN network type from isInService check
The current behavior of MobileNetworkController is to consider the
device in service if we have an out_of_service service state, but an
in_service data reg state. This is to handle "data only" SIM cards.

This change excludes IWLAN network types from this behavior because that
is the state of wifi calling and shouldn't cause the device to report
cell service.

Change-Id: I408fe282ebeb00425776b745abee0468ac0567c9
Fixes: 111958683
Test: atest com.android.settingslib.UtilsTest#isInService_voiceOutOfServiceDataOutOfService_returnFalse

Former-commit-id: 0e48aacfb339df4240ae3bb34a7bfa828a33caf5