[Magnifier-71] Adapt surface buffer to elevation
The surface created for the magnifier is larger than the magnifier
content, in order to leave space for drawing the shadow. In
I066082fb17cfb8c483c49b7011abfa9dca9de77a we made the elevation a
customizable property of Magnifier using the Builder class. However, the
buffer allocated for drawing the shadow remained a fixed value,
insufficient for large elevations, when shadows were getting clipped.
This CL fixes this, by making the allocated surface large enough to
accomodate the shadow.

Bug: none
Test: atest CtsWidgetTestCases:android.widget.cts.MagnifierTest
Change-Id: I462c50510820614dc033a42b347e9672ce39a4c6

Former-commit-id: bd391f9aa13877ebf3b6fd69f19fac058bbdd9fd