NotificationEntryListener: pre & post onEntryUpdated
Some of the listeners care about doing things to notifications prior
to the entries being filtered and some of them care about it after the
entries are filtered.

This CL introduces new method onPreEntryUpdated and alters the existing
method to be onPostEntryUpdated so that listeners can know about both
events.

Test: atest NotificationEntryManagerTest#estUpdateNotification_prePostEntryOrder
Change-Id: Ifa56fbf27cbc8f58c8eda9c95c79a57f1a52c9ac

Former-commit-id: 0ad5b9defd621c99d35f324538ee8e3af41e3cc2