Make ColorSpace#isSrgb more restrictive
Test: Ide74c2e98d4aadba8bb0a24d132f161dbe64f4db

Previously, BT709 was believed to be SRGB, resulting in errors when we
tried to convert between them. Make isSrgb test several values passed to
OETF and EOTF to verify that it is very close to SRGB. This results in
BT709 no longer being considered SRGB, fixing the conversion.

Update the documentation to reflect the new behavior.

Now that isSrgb is more restrictive, a gamma of 2.2 is no longer
considered to be SRGB, which requires a change to a CTS test.

Change-Id: I7bf8d5e3af2f91e38ca0b0d5d906713a125dd484

Former-commit-id: 203034fe2cb099f97d5ff3ea433ca657e231daef