Enable null cursor drawable
I3801ee8642d277c96c094b7d4155e28fc5bcea87 added public setters and
getter for a TextView's cursor drawable. The setter was expecting a
NonNull Drawable argument. However, this was inconsistent with the
textCursorDrawable attribute which had existed for a long time in
platform. This accepted null, case when a fallback vertical black line
is drawn as a cursor. This CL fixes this inconsistency, by allowing null
Drawables to be set for the cursor.

Also, the CL fixes the preconditions checking resource id arguments to
be strictly positive, while negative resource ids are valid as well.

Bug: 122550166
Test: atest CtsWidgetTestCases:android.widget.cts.TextViewTest
Change-Id: Ia10fafbdc0405968e57b35a615eedaeb9ee80c96

Former-commit-id: f41a054fca9ff7c0a931309ce3e47832a9044239