Fix compile time warning.
In checkIncidentPermissions(), the intention of code is to bail out when
the caller do not have appropriate permission.

Logically, DEST_LOCAL and DEST_EXPLICIT should be checked separately.
In the current code, we would fall-through to DEST_EXPLICIT case, but
because permitted DEST_LOCAL callers (AID_SHELL and AID_ROOT) are also
permitted DEST_EXPLICIT callers, we will still get the Status:ok().

Test: build
Change-Id: I0f6121cb1aba7724306b99469f5ef97b4358ac7d

Former-commit-id: 3a039ff76350811822b3ab885c64c683016da879