Enable 'real bubbles' and fix our tests
Previously the test was marking entries as 'isBubble' automatically which
isn't quite correct.

This CL fixes that and has the side effect that enables bubbles for real.
This doesn't auto-bubble any notification, however, if the notification
is on an appropriately noisy channel & has an app overlay intent we will
bubble it.

Bug: 111236845
Test: atest BubbleControllerTest
Change-Id: I26c67ce002eec808591d7864b2366bb4140be386

Former-commit-id: b4991e60dbf698bfc7460901814c3aafac7942a9