Add statsd atom PROCESS_START_TIME
The definition of app process start time is from the app
launch time to the time that Zygote finished fork the app process and
loaded the application package's java classes.

The app process can be started for any component including  activity,
service, content provider, broadcast receiver or other, either in
foreground or background.

The difference of this atom from existing atom like AppStartOccured
is that this atom can report start time for both foreground and
background process.

Bug: 122356380
Test: NA.
Change-Id: Ifcc0c80f1da45c2ff89f7c88da6407ba777473e2

Former-commit-id: 885aca1bbaa076b11914b9e4cb0d2a6d0c6fab3a