Disallow pinned stack to control occluded state.
In the original design, we check the top focus stack of each display
to control its occlusion state. However, the start time of pinned activity
may start later than the user's application, such as PipMenuActivity. In
which case the top focus activity will not be what the user expects.
To prevent this, we can ignore Pinned stack when checking visibility for
a display.

Fix: 120445909
Test: atest KeyguardTests KeyguardLockedTests
Test: atest ActivityManagerMultiDisplayTests
Test: atest ActivityManagerDisplayLockedKeyguardTests ActivityManagerDisplayKeyguardTests

Change-Id: I3be19a803aa59c21841e476aa46a0222d156e21e

Former-commit-id: 9a10e9db6ac2ba69e7e7b21c8e475acc6716d79d