Separate privapp grant for settings.intelligence
This app is moving to /product, so split it out of the grant
file on /system. For now, make it required by the /system file.
As part of moving the an app to /product, the grant file will
be made required by the app itself.

Bug: 122353713
Test: make
Exempt-From-Owner-Approval: approved already
Change-Id: I994137279b2c25938fd9967027f606a88bb39468
Merged-In: I994137279b2c25938fd9967027f606a88bb39468

Former-commit-id: 099d53b9cbe5f056e3aaa62ceeaf4845c11b354c