Schedule rollback expiration when a rollback is enabled.
Otherwise rollbacks will only be expired if the device reboots.

Currently there are no tests that scheduled rollback expiration works,
which explains why this bug wasn't found earlier. The plan is to add
tests for scheduled rollback expiration after the expiration timeout has
been made configurable, because it will be much easier to write the
tests with a configurable timeout.

Test: atest RollbackTest
Bug: 112431924
Change-Id: If16185b5db1a6f55efb83067e8f80a8bcb0816e7

Former-commit-id: babded150e6a4d61da1052910b4db6a4285b766a