[wv] Add WebView unresponsive renderer APIs.
WebViewRendererClient is a callback class responsible for informing
applications about renderer state changes.

Specifically, these callbacks added as part of this feature are:

onRendererUnresponsive(renderer) - called when the renderer associated
    with a webview does not ack an input event, or a navigation does
    not commit within 5 seconds. The callback will continue to be
    called at 5 second intervals until the renderer is killed or the
    unresponsiveness resolves. |renderer| may be null if webview is
    operating in single process mode.

onRendererResponsive(renderer) - called once when an unresponsive
    renderer becomes responsive again.

This CL also adds the ability to get a webview renderer handle from a
WebView instance, and then terminate that renderer. This allows recovery
from unresponsive renderer situations when webview is in multiprocess
mode and termination is correctly handled via
WebViewClient.onRendererUnresponsive().

Bug: 111332462
Test: CTS test TBD
Change-Id: I0ec026f539a70f412c5567dd87a86a0d851b39ae

Former-commit-id: 740b53f0f246954871a573a218cfb59d9c58b61c