Support for multi-package apex sessions.
If we have multi-package sessions, we check whether they contain APEX
packages and proceed with the pre-reboot verification of the whole
session. Session is then applied by apexd on reboot, and the state is
recovered and updated by StagingManager as soon as the sessions are
restored.

Bug: 118865310
Test: adb install-multi-package file1.apex file2.apex; verified sessions
are written correctly, restored correctly on stop/start, applied
correctly at reboot and their state is updated.
Test: atest apex_e2e_tests (to make sure we didn't regress)
Change-Id: Ia31d5badace016074d158e31c574fbad81576984

Former-commit-id: 015f935191111376b826ec0290f0a7f1240961b9