Enhance Logging When Skipping Loading Temp Rules
We only want to load a temporary rules file for processing if the
current app is debuggable or if adb is running as root.   The logic to
do this is a bit less than straightforward to read and we don't
currently log the values to help make it clearer (though we know they were
both 'false' to return early).   This change helps make the code a bit
easier to read and adds some logging that helps make reading the logs a
little easier too.

Bug: 122612037
Test: atest CtsAngleIntegrationHostTestCases
Change-Id: I38baf02ebdb847f9423a80a1093385903eda1686

Former-commit-id: 3df6643ad7652a7fb55233a2362bea757257798f