Support touch event on letterbox surface
This is part of "Compatible behavior for non-resizable activity".
It allows window to receive motion event when sliding from a letterbox
surface. That makes it easier for users to slide out the side menu of
a letterboxed activity.

Note this only works when the touchable region of the activity doesn't
cover letterbox surface (lower layer than activity).

Bug: 112288258
Test: manual - Start an activity with fixed aspect ratio and make sure
      it touchable region matches its content size, and then slide from
      letterbox region into the activity, it should receive touch event.
Change-Id: Ibff9103adbdad0ae5dd93c4a44b00f41d06c1007

Former-commit-id: 192fe76a05bd3f94ffa16f98e928a9e356dffa46