Switch fullscreen policy to use top activity instead of root
To maintain the status-quo, we need to use top activity instead
of the root activity when resolving fullscreen policy at the
task level.

The original intent behind using root activity was to reduce
jarring size-changes when users navigated between activities of
an app; however, there are apps in the wild which force one
orientation in their root activity and then force a different
orientation in subsequent activities. This caused nested
letterboxing which is obviously less desirable.

Bug: 122683367
Test: Updated TaskRecordTests. Verified that Netflix app doesn't
      have nested letterboxes.

Change-Id: I876c1e0c062b9d71412d58870d9ed917cf02ef71

Former-commit-id: 130d94fe2403dca045cb1d8aa6d52b8d0491d666