Add explicit subscription id support to DataUsageController
The existing code in DataUsageController only returns usage stats for
the default mobile data subscription. But for dual-sim, dual-standby, we
want to be able to get usage stats for other subscriptions as well, even
when they aren't currently set to be the default. So this CL adds a way
to request stats for an explicit subscription id.

Bug: 122670283
Test: make RunSettingsLibRoboTests
Change-Id: Ieb7eef512bf91292b3fb9f37b9bf589a98faa3a8

Former-commit-id: ea4cdcbc44cfbbb5d3ed4ab7d8a826bed1b70dfe