Populate cardId in resolution intent.
When LPA resolves a resolvable error, it needs to get an instance of
EuiccManager. Since we add cardId in EuiccManager from Q, we also need
to populate the cardId in the resolution intent so that when LPA solves
the resolvable errors, it uses the EuiccManage with the same cardId as
before.

Bug: 68941776
Test: test on phone
Change-Id: I780c39856c36516dbf1546eff28348a83d0a4fc9

Former-commit-id: da56f08b94f366b7e549e6ee4106b604d4f4bf1c