Fixes NotificationInfo for demoted notifications.
This change modifies NotificationInfo such that the choice to display
"alert" vs. "keep alerting" and "silence" vs. "stay silent" buttons is
based on the same criteria used to determine whether the notification is
in the "high priority" or "low priority" section of the shade.  In
particular this affects notifications whose channel is marked as high
priority but which are displayed as low priority due to demotion by an
assistant service adjustment.  NotificationMenuRow is also modified to
use the same logic for which icon to use for the button that launches
the NotificationInfo UI.

Change-Id: Id7534ca5133e51d636e41baba9404108ee216467
Fixes: 121334699
Test: atest SystemUITests, manual

Former-commit-id: caed15cd8bcacadb1761ae88a3747a7c3e130999