Use Transaction reparent to null instead of destroy for SC in DragState
When creating a drag surface for drag and drop, the system server is
responsible for destroying the surface when no longer needed. However,
the SurfaceControl was created on the client side, so when the server
calls destroy, it's not allowed since only the process that created the
surface can do it.

Instead use the Transaction reparent to null to destroy the SurfaceControl.

Change-Id: Ie1f0c230046e9928cd4b12ff7a908aad70166d64
Fixes: 122798367
Test: Drag and drop no longer keeps shadow surface on screen

Former-commit-id: b4fa081ca7b9f2828869164b4df57a79bbd5398c