Add metrics logging constants for notification location.
When logging whenever smart suggestions are visible or clicked we want
to know where their notification was shown - either as a heads-up or
in the notification shade. To be able to re-use our logging for other
events we here log the location of the notification with the smart
suggestions in it, rather than just logging whether it was shown as a
heads-up or in the shade.

Bug: 120767764
Test: N/A
Change-Id: Id32e7aacf01f6a03e2b7da2dd5923951fa8b5295

Former-commit-id: 3757d85eb868a4ed1221eca8b042c8c764e0f034