Don't over-specify testcases with verifyNoMoreInteractions.
In this specific case, there is a race condition between
verify(...).onProgramInfoChanged(...) and verifyNoMoreInteractions call.
As it's being called in the middle of the testcases, it's not possible to
pull the callback from the radio service to prevent the race condition.
However, Mockito docs ask to not overuse this method for over-specification
of the tests, so let's just comply with what the doc says.

Bug: 122715569
Test: atest BroadcastRadioTests
Change-Id: I978b13c7c7607773f12cc5a3a9f72f782fb3f345

Former-commit-id: f54fc26af70d5ee42d7c67398e8fb645a80d8dd5