Post expensive ContentCapture calls to Choreographer's CALLBACK_COMMIT stage.
onProvideContentCaptureStructure() doesn't take much for only 1 view, but it's called for all views
the first time the view hiearchy is laid off, which could acccumulative delay the
initial layout. Hence, we're postponing it to a later stage - it might still cost a
lost frame (or more), but that jank cost would only happen after the 1st layout.

Using this change, the impact on cold-dropcache-test dropped about 50% (from ~2ms to ~1ms).

Bug: 123307965
Bug: 121039624
Test: atest ContentCaptureSession
Test: atest google/perf/app-startup/benchmark-app-hermetic/cold-dropcache-test

Change-Id: I68b98b2894d23309af90d87cc99280f133557252

Former-commit-id: 26cbd718964d2a5474e97e5385e733452283cec4