Multiple resumed activities
This is the first step to connecting the visibility updates to
activity lifecycle - each time when an activity is made visible,
we check if it should be made active (resume or paused) and report
the corresponding state to client if needed.

The main logical change is in
ActivityStack#ensureActivitiesVisibleLocked. The idea is that calling
ActivityRecord#makeActiveIfNeeded() should identify what should the
state of an activity be and and perform the transition. This should
eventually reduce the number of different places where we trigger
lifecycle transitions.

Now all top activities in visible and focusable stacks should be
resumed.

Bug: 117135575
Bug: 123166999
Test: atest android.server.am.lifecycle
Test: atest WmTests:ActivityRecordTests
Change-Id: I4954c4fa838c1a488892d7447e9ac1e48086c221

Former-commit-id: 6b321517480924e5e977e0ef2437457f384280ab