WifiConfiguration: App attribution for suggestion/request
Add new field to mark a WifiConfiguration as created from a network
specifier or suggestion. WifiConfiguration.creatorName for such networks
will contain the package name of the app that added the corresponding
specifier/suggestion.
This can be used by the settings app to display the app's name in the
wifi picker summary.

Bug: 115504887
Bug: 113878056
Test: ./frameworks/base/wifi/tests/runtests.sh
Change-Id: Ic0ae2292930628dc38c668777c7de4f19af4e925

Former-commit-id: f15a9720faf3fc6fbd3da9c1632fab4155c319fd