Migrate to single sanely-sized thumbnails.
Instead of generating only a handful of thumbnail "kinds", none of
which scale with display density, shift to generating only a single
thumbnail whose dimensions are limited to half the smallest screen
width of the device.

This change also shifts to storing thumbnails on the relevant volume
where the source data actually came from.  This solves the problem
of having to GC thumbnails for a storage device that we haven't seen
in a long time.

Bug: 119887587
Test: atest cts/tests/tests/provider/src/android/provider/cts/MediaStore*
Change-Id: I2ce6c5051b7058d2f1bba523c077e84a1f0b46e9

Former-commit-id: 36ee36238c48dd13dea2242304696e7058f03f6b