Deprecate InputMethodManager#setCurrentInputMethodSubtype()
InputMethodManager#setCurrentInputMethodSubtype(), which was
introduced in Android 4.0 ICS [1], was probably mistakenly exposed as
a public API, because it has required WRITE_SECURE_SETTINGS that
typical applications cannot have.

This CL marks InputMethodManager#setCurrentInputMethodSubtype()
deprecated.  There is no actual implementation change at this moment,
but developers should avoid relying on this API in general not only on
futuer versions of Android but also on existing devices.

 [1]: I55daa19ba924999def544bf841f00bf54852f3e1
      b66d287e3003a0934d5714fbf15e554b3c814906

Fix: 123249820
Test: make -j checkbuild
Change-Id: Ia31df87ad2182484d7335503c5d87cb93540fd17

Former-commit-id: ff42b1b6853b631d294796c9b118686d0d58b408